requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
configparser==5.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
Cython==0.29.33
defusedxml==0.7.1
Django==4.1.7
django-environ==0.10.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
numpy==1.24.2
oauthlib==3.2.2
pandas==1.5.3
Pillow==9.5.0
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.1
sqlparse==0.4.3
Telethon==1.27.0
uritemplate==4.1.1
urllib3==1.26.14
